Arithmetic 타입변환, Arrange class 생성, Calculator 마무리
Ouput 출력

+++ Calculator/src/com/calculator/base/Arithmetic.java
@@ -1,8 +1,8 @@
 package com.calculator.base;
 
 public interface Arithmetic {
-	public int add(int num1, int num2);
-	public int sub(int num1, int num2);
-	public int mul(int num1, int num2);
-	public double div(int num1, int num2);
+	public double add(double num1, double num2);
+	public double sub(double num1, double num2);
+	public double mul(double num1, double num2);
+	public double div(double num1, double num2);
 }
